Extract Prompt Studio advanced field state helpers
